requirements.txt
schedule
requests
beautifulsoup4
gTTS
pandas
openpyxl # Needed by pandas to write .xlsx files
python-dotenv
transformers>=4.38.0 # Use a recent version
torch>=2.1.0 # Required by transformers
accelerate>=0.25.0 # Recommended for model loading/performance
# bitsandbytes>=0.41.0 # Optional: For 8-bit/4-bit quantization (if needed for memory) - Linux/Windows only usually
# sentencepiece # Often needed by tokenizers

bitsandbytes #  enable 4-bit quantization for limited memory (loads the model using lower precision (like 4-bit instead of 16-bit or 32-bit), significantly reducing the memory)

numpy<2.0 # Specify NumPy version constraint

#llama-cpp-python>=0.2.18 # Or a specific newer version if desired

#CMAKE_ARGS="-DLLAMA_METAL=on" pip install --force-reinstall --no-cache-dir llama-cpp-python -> run in terminal
#(This tells the installer to compile the Metal-specific code).

